+++ internal/sensors/dht22.go
@@ -2,6 +2,6 @@ package sensors
 
 type DHT22Sensor struct {
 	Mac         string  `json:"mac"`
-	Humidity    float32 `json:"humidity"`
-	Temperature float32 `json:"temp"`
+	Humidity    float64 `json:"humidity"`
+	Temperature float64 `json:"temp"`
 }
